Adds weather icons. Implements the forecast view controller, forecast tableview cell, fleshes out the view model and makes some minor tweaks here and there.

+++ Windy City/Model/Forecast.swift
@@ -37,7 +37,7 @@ struct Forecast : Codable {
   }
   
   struct ThreeHourForecast : Codable {
-    var dt: Int64
+    var dt: Double
     var main: Main
     var weather : Array<Weather>
     var wind : Wind
